Language_update
It is almost time to migrate to the pi!

I have made hardware changes.

I am using Bash for installation processes.

I am using a Nano as an RTOS to control robot actions, and a pi as a GUOS to do the heavy lifting for the AI and communication between the Pi  and the Nano.